feat : pieces 문자열을 도메인 객체로 변경해주는 기능을 PiecesConverter 객체로 분리
